+++ CHANGES.md
@@ -10,6 +10,8 @@ Changes in CUPS v2.2.13
 - Fixed the default common name for TLS certificates used by `ippserver`.
 - The libusb-based USB backend now reports an error when the distribution
   permissions are wrong (Issue #5658)
+- The IPP backend did not detect all cases where a job should be retried using
+  a raster format (rdar://56021091)
 
 
 Changes in CUPS v2.2.12